Split Vert.x utilities to VertxUtil
