+++ operator-common/src/test/java/io/strimzi/operator/common/operator/resource/AbstractNonNamespacedResourceOperatorIT.java
@@ -10,7 +10,7 @@ import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientBuilder;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.strimzi.operator.common.Reconciliation;
-import io.strimzi.operator.common.Util;
+import io.strimzi.operator.common.VertxUtil;
 import io.strimzi.test.k8s.cluster.KubeCluster;
 import io.vertx.core.Vertx;
 import io.vertx.core.WorkerExecutor;
@@ -98,7 +98,7 @@ public abstract class AbstractNonNamespacedResourceOperatorIT<C extends Kubernet
             .onComplete(context.succeeding(rrDelete -> context.verify(() -> {
                 // it seems the resource is cached for some time so we need wait for it to be null
                 context.verify(() -> {
-                        Util.waitFor(Reconciliation.DUMMY_RECONCILIATION, vertx, "resource deletion " + resourceName, "deleted", 1000,
+                        VertxUtil.waitFor(Reconciliation.DUMMY_RECONCILIATION, vertx, "resource deletion " + resourceName, "deleted", 1000,
                                 30_000, () -> op.get(resourceName) == null)
                                 .onComplete(del -> {
                                     assertThat(op.get(resourceName), is(nullValue()));
